# Backend/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "ProjectHub",
    "description": "Jesus Antona Espejo",
    "version": "0.1.0"
  },
  "paths": {
    "/Projects/": {
      "get": {
        "tags": [
          "Projects"
        ],
        "summary": "Find All Projects",
        "description": "Devuelve una lista de proyectos",
        "operationId": "find_all_projects_Projects__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Project"
                  },
                  "type": "array",
                  "title": "Response Find All Projects Projects  Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Projects"
        ],
        "summary": "Create Project",
        "description": "Crea un project y lo devuelve",
        "operationId": "create_project_Projects__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Project"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Project"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/Projects/{id}": {
      "get": {
        "tags": [
          "Projects"
        ],
        "summary": "Find One Project",
        "description": "Devuelve el project con id pasado por par\u00e1metro",
        "operationId": "find_one_project_Projects__id__get",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Id del project a buscar",
              "title": "Id"
            },
            "description": "Id del project a buscar"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Project"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Projects"
        ],
        "summary": "Delete Project",
        "description": "Elimina el project con id pasado por par\u00e1metro",
        "operationId": "delete_project_Projects__id__delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Id del project a eliminar",
              "title": "Id"
            },
            "description": "Id del project a eliminar"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Projects"
        ],
        "summary": "Update Project",
        "description": "Actualiza el project con id pasado por par\u00e1metro y lo devuelve",
        "operationId": "update_project_Projects__id__put",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Id del project a actualizar",
              "title": "Id"
            },
            "description": "Id del project a actualizar"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/Project"
                  }
                ],
                "description": "Datos del project a actualizar",
                "title": "Project"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Project"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Project": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "company": {
            "type": "string",
            "title": "Company"
          },
          "parent_id": {
            "type": "string",
            "title": "Parent Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "planned_value": {
            "type": "number",
            "title": "Planned Value"
          },
          "actual_cost": {
            "type": "number",
            "title": "Actual Cost"
          },
          "start_date": {
            "type": "string",
            "title": "Start Date"
          },
          "finish_date": {
            "type": "string",
            "title": "Finish Date"
          },
          "start_real_date": {
            "type": "string",
            "title": "Start Real Date"
          },
          "finish_real_date": {
            "type": "string",
            "title": "Finish Real Date"
          },
          "risk": {
            "type": "number",
            "title": "Risk"
          },
          "priority": {
            "type": "string",
            "title": "Priority"
          },
          "strategic_goal": {
            "type": "string",
            "title": "Strategic Goal"
          },
          "earned_value": {
            "type": "number",
            "title": "Earned Value"
          }
        },
        "type": "object",
        "required": [
          "code",
          "company",
          "parent_id",
          "title",
          "description",
          "planned_value",
          "actual_cost",
          "start_date",
          "finish_date",
          "start_real_date",
          "finish_real_date",
          "risk",
          "priority",
          "strategic_goal",
          "earned_value"
        ],
        "title": "Project"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  },
  "servers": [
    {
      "url": "/",
      "description": "Default"
    },
    {
      "url": "http://localhost:8000",
      "description": "Local dev"
    }
  ]
}